initial_mask2former/requirements.txt
numpy<2.0.0
jupyterlab
torch>=1.8.0
ipykernel
tqdm
ipywidgets
Pillow

# OpenMMLab Core
openmim
mmengine>=0.7.4
mmcv>=2.0.0
mmdet>=3.0.0
mmsegmentation>=1.2.2

# Mask2Former Specifics
scipy
ftfy
regex

# Visualization & Utils
matplotlib
opencv-python
seaborn
tensorboardX
future 
tensorboard